refactor(blog): use dash case filename
